Fixed Position Player. #30
Fixed position player controls/info box.
I don't know how to do it dynamically, so it has a 50px margin from
playlist. This works great on big screens, but if you have a long file
path on a small screen it can have some over lap with the playlist.

Either need to figure out how to position the playlist dynamically or
switch the the file name to a one line ticker. (Which I may do, since
I'd prefer to keep the height of the controls to a minimum.)